fix rpm preparer
