+++ resource/packages/rpm/rpm.go
@@ -19,8 +19,8 @@ import (
 	"os/exec"
 	"strings"
 
-	log "github.com/Sirupsen/logrus"
 	"github.com/asteris-llc/converge/resource"
+	"github.com/pkg/errors"
 )
 
 // Package is an API for package state
@@ -28,30 +28,44 @@ type Package struct {
 	resource.TaskStatus
 
 	Name  string
-	State string
+	State State
 }
+
+// State type for Package
+type State string
+
+const (
+	// StatePresent indicates the package should be present
+	StatePresent State = "present"
+
+	// StateAbsent indicates the package should be absent
+	StateAbsent State = "absent"
+)
 
 // Check if the package has to be 'installed', or 'absent'
 func (p *Package) Check(resource.Renderer) (resource.TaskStatus, error) {
 	status := resource.NewStatus()
 
-	currentPkgStatusRaw, err := exec.Command("bash", "-c", fmt.Sprintf("yum info %q | grep Repo | awk '{print $3}'", p.Name)).Output()
+	currentPkgStatusRaw, err := exec.Command("sh", "-c", fmt.Sprintf("yum info %q | grep Repo | awk '{print $3}'", p.Name)).Output()
 	if err != nil {
-		log.WithField("module", "rpm").Warn(err)
+		return status, errors.Wrapf(err, "checking package %s", p.Name)
 	}
 	currentPkgStatus := strings.TrimSpace(string(currentPkgStatusRaw))
 
-	log.WithField("module", "rpm").WithField("exec", "bash").Info(string(currentPkgStatus))
-	log.Info(p)
-	if p.State == "installed" {
+	switch p.State {
+	case StatePresent:
 		if string(currentPkgStatus) != "installed" {
 			status.AddDifference(p.Name, string(currentPkgStatus), "installed", "")
+			status.RaiseLevel(resource.StatusWillChange)
 		} else {
 			status.AddMessage("Package is installed")
 		}
-	} else if p.State == "absent" {
+	case StateAbsent:
 		if string(currentPkgStatus) == "installed" {
 			status.AddDifference(p.Name, string(currentPkgStatus), "uninstalled", "")
+			status.RaiseLevel(resource.StatusWillChange)
+		} else {
+			status.AddMessage("Package is absent")
 		}
 	}
 
@@ -63,31 +77,26 @@ func (p *Package) Check(resource.Renderer) (resource.TaskStatus, error) {
 func (p *Package) Apply() (resource.TaskStatus, error) {
 	status := resource.NewStatus()
 
-	if p.State == "installed" {
-		currentPkgStatus, err := exec.Command("bash", "-c", "yum install -y "+p.Name).Output()
+	switch p.State {
+	case StatePresent:
+		currentPkgStatus, err := exec.Command("sh", "-c", "yum install -y "+p.Name).Output()
+
 		if err != nil {
-			log.WithField("module", "rpm").Warn(err)
-			log.WithField("module", "rpm").Warn(string(currentPkgStatus))
+			return status, errors.Wrapf(err, "installing package %s, what happened: %s", p.Name, currentPkgStatus)
 		}
 		status.AddMessage(fmt.Sprintf("Package %q installed", p.Name))
 		status.AddDifference(p.Name, "absent", "installed", "")
 
-		log.WithField("module", "rpm").Info(fmt.Sprintf("Installed %q: \n %q", p.Name, string(currentPkgStatus)))
-	} else if p.State == "absent" {
-		currentPkgStatus, err := exec.Command("bash", "-c", "yum remove -y "+p.Name).Output()
-		if err != nil {
-			log.WithField("module", "rpm").Warn(err)
-			log.WithField("module", "rpm").Warn(string(currentPkgStatus))
-		}
+	case StateAbsent:
+		currentPkgStatus, err := exec.Command("sh", "-c", "yum remove -y "+p.Name).Output()
 
+		if err != nil {
+			return status, errors.Wrapf(err, "installing package %s, what happened: %s", p.Name, currentPkgStatus)
+		}
 		status.AddMessage(fmt.Sprintf("Package %q removed", p.Name))
 		status.AddDifference(p.Name, "installed", "absent", "")
-
-		log.WithField("module", "rpm").Info(fmt.Sprintf("Removed %q: \n %q", p.Name, string(currentPkgStatus)))
 	}
 
-	status.RaiseLevel(resource.StatusNoChange)
 	p.TaskStatus = status
-
 	return p, nil
 }
